Feature/#65 item yuchan

See merge request cloud_track/class_04/web_project2/team06/Shoux-Kream!63

+++ src/main/java/com/shoux_kream/item/dto/ItemDto.java
@@ -31,10 +31,10 @@ public class ItemDto {
     private String imageKey;
 
     @Positive(message = "재고 수량은 양수여야 합니다.")
-    private int inventory;
+    private Integer inventory;
 
     @Positive(message = "가격은 양수여야 합니다.")
-    private double price;
+    private Integer price;
 
     private String searchKeywords;
 }
